+++ sidertests/__init__.py
@@ -1,7 +1,8 @@
 import doctest
 import os
 from attest import Tests
-from . import session, types, hash, list, set
+from . import (session, types, hash, list, set, sortedset,
+               transaction, threadlocal)
 
 
 tests = Tests()
@@ -10,6 +11,9 @@ tests.register(types.tests)
 tests.register(hash.tests)
 tests.register(list.tests)
 tests.register(set.tests)
+tests.register(sortedset.tests)
+tests.register(transaction.tests)
+tests.register(threadlocal.tests)
 
 
 @tests.test