+++ docs/changes.rst
@@ -1,6 +1,16 @@
 Sider Changelog
 ===============
 
+Version 0.2.0
+-------------
+
+- Added :mod:`sider.transaction` module.
+- Added :mod:`sider.sortedset` module.
+- Added :class:`sider.types.SortedSet` type.
+- Added :mod:`sider.threadlocal` module.
+- Added :attr:`sider.session.Session.verbose_transaction_error` option.
+
+
 Version 0.1.3
 -------------